refactor(loader): remove legacy task-detection API; relocate helpers to resolution
