+++ src/winml/modelkit/eval/evaluate.py
@@ -228,7 +228,7 @@ def _resolve_task(config: WinMLEvaluationConfig) -> str:
     """Resolve the eval task and validate it is supported.
 
     An explicit ``config.task`` is surfaced verbatim (explicit means explicit).
-    When omitted, the modality-aware :func:`detect_task` infers it from the model's
+    When omitted, the modality-aware :func:`resolve_task` infers it from the model's
     HF config — an image-embedding model resolves to ``image-feature-extraction``
     (not the lossy ``feature-extraction``), so the evaluator-registry lookup picks
     the image evaluator without any reverse io_config reconstruction.
