+++ src/winml/modelkit/loader/__init__.py
@@ -26,16 +26,15 @@ Note:
 """
 
 from .config import WinMLLoaderConfig, resolve_loader_config
+from .resolution import TaskResolution, TaskSource, resolve_composite, resolve_task
 from .task import (
     HF_TASK_DEFAULTS,
     KNOWN_TASKS,
     TASK_SYNONYM_EXTENSIONS,
-    detect_task,
     get_supported_tasks,
     get_task_abbrev,
     normalize_task,
     resolve_optimum_library,
-    resolve_task_and_model_class,
     to_optimum_task,
 )
 
@@ -44,16 +43,18 @@ __all__ = [
     "HF_TASK_DEFAULTS",
     "KNOWN_TASKS",
     "TASK_SYNONYM_EXTENSIONS",
+    "TaskResolution",
+    "TaskSource",
     "WinMLLoaderConfig",
-    "detect_task",
     "get_supported_tasks",
     "get_task_abbrev",
     "load_hf_model",
     "normalize_task",
+    "resolve_composite",
     "resolve_hf_model_class",
     "resolve_loader_config",
     "resolve_optimum_library",
-    "resolve_task_and_model_class",
+    "resolve_task",
     "to_optimum_task",
 ]
