+++ tests/integration/loader/test_hf_model_class_mapping.py
@@ -47,7 +47,7 @@ class TestHFModelClassMappingE2E:
 
 
 class TestResolveTaskAndModelClass:
-    """Tests for the resolve_task_and_model_class() function.
+    """Tests for the resolve_task() function.
 
     Moved from tests/unit/ — these tests call AutoConfig.from_pretrained()
     which downloads model configs from HuggingFace Hub.
@@ -57,125 +57,125 @@ class TestResolveTaskAndModelClass:
         """Case 1: task=None, model_class=None → auto-detect both."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("microsoft/resnet-18")
-        task, resolved_class = resolve_task_and_model_class(config)
+        r = resolve_task(config)
 
-        assert task == "image-classification"
-        assert "ImageClassification" in resolved_class.__name__
+        assert r.task == "image-classification"
+        assert "ImageClassification" in r.model_class.__name__
 
     def test_case2_task_only_standard(self):
         """Case 2: task specified, model_class=None → resolve for task."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("microsoft/resnet-18")
-        task, resolved_class = resolve_task_and_model_class(config, task="image-classification")
+        r = resolve_task(config, task="image-classification")
 
-        assert task == "image-classification"
-        assert "ImageClassification" in resolved_class.__name__
+        assert r.task == "image-classification"
+        assert "ImageClassification" in r.model_class.__name__
 
     def test_case2_task_only_with_specialization(self):
         """Case 2: task specified triggers specialization (CLIP)."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("openai/clip-vit-base-patch32")
 
         # image-feature-extraction should return CLIPVisionModelWithProjection
-        task, resolved_class = resolve_task_and_model_class(config, task="image-feature-extraction")
-        assert task == "image-feature-extraction"  # preserves original task name
-        assert resolved_class.__name__ == "CLIPVisionModelWithProjection"
+        r = resolve_task(config, task="image-feature-extraction")
+        assert r.task == "image-feature-extraction"  # preserves original task name
+        assert r.model_class.__name__ == "CLIPVisionModelWithProjection"
 
         # feature-extraction should return CLIPTextModelWithProjection directly
-        task, resolved_class = resolve_task_and_model_class(config, task="feature-extraction")
-        assert task == "feature-extraction"
-        assert resolved_class.__name__ == "CLIPTextModelWithProjection"
+        r = resolve_task(config, task="feature-extraction")
+        assert r.task == "feature-extraction"
+        assert r.model_class.__name__ == "CLIPTextModelWithProjection"
 
     def test_case2_task_only_with_segformer_specialization(self):
         """Case 2: task specified triggers specialization (Segformer)."""
         from transformers import SegformerConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = SegformerConfig()
 
-        task, resolved_class = resolve_task_and_model_class(config, task="image-segmentation")
-        assert task == "image-segmentation"
-        assert resolved_class.__name__ == "AutoModelForSemanticSegmentation"
+        r = resolve_task(config, task="image-segmentation")
+        assert r.task == "image-segmentation"
+        assert r.model_class.__name__ == "AutoModelForSemanticSegmentation"
 
     def test_case2_task_only_nsp(self):
         """Case 2: NSP task uses HF_TASK_DEFAULTS."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("prajjwal1/bert-tiny")
-        task, resolved_class = resolve_task_and_model_class(config, task="next-sentence-prediction")
+        r = resolve_task(config, task="next-sentence-prediction")
 
-        assert task == "next-sentence-prediction"
+        assert r.task == "next-sentence-prediction"
         # AutoModelForNextSentencePrediction resolves to concrete class
-        assert "NextSentencePrediction" in resolved_class.__name__
+        assert "NextSentencePrediction" in r.model_class.__name__
 
     def test_case3_model_class_override(self):
         """Case 3: model_class specified → honor it."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("openai/clip-vit-base-patch32")
 
         # Explicitly request CLIPModel even though specialization would pick something else
-        task, resolved_class = resolve_task_and_model_class(
+        r = resolve_task(
             config,
             task="feature-extraction",
             model_class="CLIPModel",
         )
 
-        assert task == "feature-extraction"
-        assert resolved_class.__name__ == "CLIPModel"
+        assert r.task == "feature-extraction"
+        assert r.model_class.__name__ == "CLIPModel"
 
     def test_case3_model_class_with_auto_task(self):
         """Case 3: model_class with task=None → detect task."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("microsoft/resnet-18")
 
         # Specify model_class but let task be auto-detected
-        task, resolved_class = resolve_task_and_model_class(
+        r = resolve_task(
             config,
             model_class="AutoModelForImageClassification",
         )
 
-        assert task == "image-classification"
-        assert "ImageClassification" in resolved_class.__name__
+        assert r.task == "image-classification"
+        assert "ImageClassification" in r.model_class.__name__
 
     def test_invalid_task_raises_error(self):
         """Invalid task should raise ValueError."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("microsoft/resnet-18")
 
         with pytest.raises(ValueError, match="not supported"):
-            resolve_task_and_model_class(config, task="invalid-nonexistent-task")
+            resolve_task(config, task="invalid-nonexistent-task")
 
     def test_invalid_model_class_raises_error(self):
         """Invalid model_class should raise ValueError."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("microsoft/resnet-18")
 
         with pytest.raises(ValueError, match="not found"):
-            resolve_task_and_model_class(
+            resolve_task(
                 config,
                 task="image-classification",
                 model_class="NonExistentModel",
@@ -203,21 +203,21 @@ class TestConflictScenarios:
         """
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("openai/clip-vit-base-patch32")
 
         # Without model_class: specialization picks CLIPTextModelWithProjection
-        _task, resolved_class = resolve_task_and_model_class(config, task="feature-extraction")
-        assert resolved_class.__name__ == "CLIPTextModelWithProjection"
+        r = resolve_task(config, task="feature-extraction")
+        assert r.model_class.__name__ == "CLIPTextModelWithProjection"
 
         # With model_class: user override takes precedence
-        _task, resolved_class = resolve_task_and_model_class(
+        r = resolve_task(
             config,
             task="feature-extraction",
             model_class="CLIPModel",
         )
-        assert resolved_class.__name__ == "CLIPModel"
+        assert r.model_class.__name__ == "CLIPModel"
 
     def test_mismatched_task_model_class_honored(self):
         """model_class is honored even if mismatched with task.
@@ -227,20 +227,20 @@ class TestConflictScenarios:
         """
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("microsoft/resnet-18")
 
         # Request image-classification task but with feature extraction model
         # TasksManager validates this is a legal combination
-        task, resolved_class = resolve_task_and_model_class(
+        r = resolve_task(
             config,
             task="image-classification",
             model_class="AutoModelForImageClassification",
         )
 
-        assert task == "image-classification"
-        assert "ImageClassification" in resolved_class.__name__
+        assert r.task == "image-classification"
+        assert "ImageClassification" in r.model_class.__name__
 
     def test_task_normalizes_before_specialization_lookup(self):
         """Task should be normalized before checking specializations.
@@ -250,17 +250,17 @@ class TestConflictScenarios:
         """
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("openai/clip-vit-base-patch32")
 
         # Original task "image-feature-extraction" should find specialization
-        task, resolved_class = resolve_task_and_model_class(config, task="image-feature-extraction")
+        r = resolve_task(config, task="image-feature-extraction")
 
         # Task preserves original name (normalization is internal for lookup)
-        assert task == "image-feature-extraction"
+        assert r.task == "image-feature-extraction"
         # Specialization uses original to pick Vision model
-        assert resolved_class.__name__ == "CLIPVisionModelWithProjection"
+        assert r.model_class.__name__ == "CLIPVisionModelWithProjection"
 
     def test_specialization_not_found_falls_back_to_default(self):
         """When specialization not found, fall back to TasksManager default.
@@ -270,7 +270,7 @@ class TestConflictScenarios:
         """
         from transformers import AutoConfig
 
-        from winml.modelkit.loader.task import _get_custom_model_class
+        from winml.modelkit.loader.resolution import _get_custom_model_class
 
         # BERT doesn't have specialization for feature-extraction
         bert_config = AutoConfig.from_pretrained("prajjwal1/bert-tiny")
@@ -283,15 +283,15 @@ class TestConflictScenarios:
         """model_class with task=None should auto-detect task."""
         from transformers import AutoConfig
 
-        from winml.modelkit.loader import resolve_task_and_model_class
+        from winml.modelkit.loader import resolve_task
 
         config = AutoConfig.from_pretrained("microsoft/resnet-18")
 
-        task, resolved_class = resolve_task_and_model_class(
+        r = resolve_task(
             config,
             model_class="AutoModelForImageClassification",
         )
 
         # Task auto-detected from config.architectures
-        assert task == "image-classification"
-        assert "ImageClassification" in resolved_class.__name__
+        assert r.task == "image-classification"
+        assert "ImageClassification" in r.model_class.__name__
